Переработал util классы для генерации фото у сообщений. Покрыл их комментариями, изменил названия, вынес часть логики в класс SpecificMediaMessageCreator. Разбил util классы на отдельные внутренние папки.
